+++ src/accounts/controllers/accounts.controller.ts
@@ -1,14 +1,6 @@
-import {
-  Body,
-  Controller,
-  Post,
-  Get,
-  UseGuards,
-  Put,
-  Request,
-} from '@nestjs/common';
+import { Body, Controller, Get, UseGuards, Put, Request } from '@nestjs/common';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
-import { CreateAccountDto, UpdateAccountDto } from '../dtos/accounts.dto';
+import { UpdateAccountDto } from '../dtos/accounts.dto';
 import { AccountsService } from '../services/accounts.service';
 
 @UseGuards(JwtAuthGuard)
@@ -20,12 +12,6 @@ export class AccountsController {
   findByUser(@Request() req) {
     const userId = req.user.sub;
     return this.accountsService.findByUser(userId);
-  }
-
-  @Post()
-  create(@Body() payload: CreateAccountDto, @Request() req) {
-    const userId = req.user.sub;
-    return this.accountsService.createOneAccount(payload, userId);
   }
 
   @Put()
